Extract CartViewModel to own file
